Revert "Text, Event, Start-up. Refactored, Template, Builders. Fixed attack"

+++ core/game/src/main/kotlin/castle/core/game/system/HPSystem.kt
@@ -3,10 +3,14 @@ package castle.core.game.system
 import castle.core.common.`object`.CommonEntity
 import castle.core.common.component.PositionComponent
 import castle.core.common.config.GUIConfig
+import castle.core.common.system.IteratingIntervalSystem
+import castle.core.game.component.BehaviourComponent
 import castle.core.game.component.HPComponent
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.core.Family
 import com.badlogic.ashley.systems.IteratingSystem
+import com.badlogic.gdx.graphics.g2d.Animation
+import com.badlogic.gdx.graphics.g2d.TextureRegion
 import com.badlogic.gdx.graphics.g3d.decals.DecalBatch
 import com.badlogic.gdx.math.Vector3
 
@@ -22,7 +26,7 @@ class HPSystem(guiConfig: GUIConfig) : IteratingSystem(Family.all(PositionCompon
     override fun processEntity(entity: Entity, deltaTime: Float) {
         val hpComponent = HPComponent.mapper.get(entity)
         hpComponent.decal.position = hpComponent.matrix4.getTranslation(temp).add(hpComponent.translation)
-        val percent = hpComponent.currentAmount.toFloat() / hpComponent.amount.toFloat()
+        val percent = hpComponent.currentAmount.toFloat() / hpComponent.amount
         hpComponent.decal.textureRegion = hpComponent.animation.getKeyFrame(hpComponent.animation.animationDuration * percent)
         decalBatch.add(hpComponent.decal)
         if (hpComponent.currentAmount <= 0) {
